requirements.txt
llama-index>=0.10.20
llama-index-vector-stores-qdrant>=0.5.0
openai>=1.3.0
qdrant-client>=1.6.1
sentence-transformers>=2.2.2
python-dotenv>=0.9.9
tqdm>=4.65.0
markdown>=3.4.0
beautifulsoup4>=4.12.2